backendModel/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-dotenv
gdown
opencv-python-headless
